Downlaod file if force

+++ functions/initialize.sh
@@ -110,7 +110,7 @@ function download()
     debug_log "Downloading file $2 => $1"
     wget --no-clobber --quiet --no-cache --no-cookies -O $1 $2
   fi
-  if [ -f $1 ] && [ "$(cat $1 | wc -l)" -gt "0" ] || [ "$3" != "-force" ]
+  if [ -f $1 ] && [ "$(cat $1 | wc -l)" -gt "0" ] && [ "$3" != "-force" ]
   then
     log_bold "Skipping file as it exists $2 to $1 with flag $3"
     echo > /dev/null
